rfq: move price oracle mock to mock.go
This change aligns with the existing mock organization pattern used in
other packages.
